Reorganize dirctory for Header component
