+++ src/components/Rotues/index.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import CryptoQuote from '../../containers/CryptoQuote';
 import Trending from '../../containers/Trending';
 import App from '../App';
-import Header from '../../containers/Header';
+import Header from '../Header';
 import NotFound from '../NotFound';
 import Footer from '../Footer';
